+++ types/cfxaddress/address_test.go
@@ -6,6 +6,7 @@ import (
 	"testing"
 
 	"github.com/ethereum/go-ethereum/rlp"
+	"gotest.tools/assert"
 )
 
 func TestCfxAddress(t *testing.T) {
@@ -221,4 +222,10 @@ func fatalIfErr(t *testing.T, err error) {
 
 func GetAddressPtr(addr Address) *Address {
 	return &addr
+}
+
+func TestCfxMappedEVMSpaceAddress(t *testing.T) {
+	cfxAddr := MustNewFromBase32("cfx:aak2rra2njvd77ezwjvx04kkds9fzagfe6ku8scz91")
+	evmAddr, _ := cfxAddr.GetMappedEVMSpaceAddress()
+	assert.Equal(t, "0x12Bf6283CcF8Ad6ffA63f7Da63EDc217228d839A", evmAddr.String())
 }
